+++ src/main/scala/mada/peg/AndIf.scala
@@ -12,7 +12,7 @@ object AndIf {
 }
 
 class AndIfPeg[A](override val self: Peg[A], pred: Vector.Func3[A, Boolean]) extends PegProxy[A] {
-    override def parse(v: Vector[A], first: Long, last: Long) = {
+    override def parse(v: Vector[A], first: Int, last: Int) = {
         val cur = self.parse(v, first, last)
         if (cur == Peg.FAILURE || !pred(v, first, cur)) {
             Peg.FAILURE
