+++ src/main/scala/mada/peg/ByNeedActions.scala
@@ -18,7 +18,7 @@ class ByNeedActions[A] {
     def need(p: Peg[A]): Peg[A] = new NeedPeg(p)
 
     class NeedPeg(override val self: Peg[A]) extends PegProxy[A] {
-        override def parse(v: Vector[A], first: Long, last: Long) = {
+        override def parse(v: Vector[A], first: Int, last: Int) = {
             val cur = self.parse(v, first, last)
             if (cur != Peg.FAILURE) {
                 fireActions
