+++ src/main/scala/mada/peg/StarUntil.scala
@@ -12,7 +12,7 @@ object StarUntil {
 }
 
 class StarUntilPeg[A](p: Peg[A], q: Peg[A]) extends Peg[A] {
-    override def parse(v: Vector[A], first: Long, last: Long): Long = {
+    override def parse(v: Vector[A], first: Int, last: Int): Int = {
         var cur = first
 
         var next = q.parse(v, cur, last)
